Don't print loaded message for each module
There's a lot of modules in this bot now.
Since we show how many triggers, help responses etc. are registered, and
this gives us the same information in a more compact space, we might as
well just rely on that.

Besides, this will make it easier to see modules that choose not to load
(e.g. those requiring SQL when it is unavailable)